Add additional checks and port number option
The $_GET commands are now checked with isset() to prevent warnings from showing up. They only would if you had display_errors enabled, but it's still good to fix.

The $host variable is now checked to see if it is separated with a colon, indicating that it has a port number included. If it is included, a connection will be made to the database using the desired port number.

Indentation has been fixed so everything is more readable.